Moved debug crash dialog to `zng-wgt-inspector` crate.
